Improve basic_formula so you can make f12/2 alone
previously it was possible to do f32+f12/2 but not f12/2.
this was really weird from the user side so this commit tries to correct things. Basic testing was done
Dont know why I did not go into that solution first so I am still not 100% sure. should be tested by users at OEM forum